requirements.txt
flask
flask-cors
python-dotenv
pydantic
httpx